refactor: 컨밴션 리팩토링
