+++ src/test/java/shopping/intergration/LoginIntegrationTest.java
@@ -13,8 +13,8 @@ import shopping.exception.ErrorCode;
 import static org.assertj.core.api.Assertions.assertThat;
 import static shopping.intergration.helper.LogInHelper.loginRequest;
 import static shopping.intergration.helper.RestAssuredHelper.extractObject;
-import static shopping.intergration.util.LoginUtils.EMAIL;
-import static shopping.intergration.util.LoginUtils.PASSWORD;
+import static shopping.intergration.utils.LoginUtils.EMAIL;
+import static shopping.intergration.utils.LoginUtils.PASSWORD;
 
 class LoginIntegrationTest extends IntegrationTest {
